### 2024-10-18 #7370 @editorialbot
[PRE REVIEW]: QuantumACES.jl: Design noise characterisation experiments for quantum computers
**Submitting author:** @evanhockings (<a href="http://orcid.org/0000-0002-5890-5705">Evan T. Hockings</a>)
**Repository:** https://github.com/evanhockings/QuantumACES.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
