### 2024-11-20 #7370 @phibeck
Hi @evanhockings, thanks for your submission. I'll be looking for reviewers next. If you have a moment it would be helpful if you could identify a few potential reviewers from this list https://reviewers.joss.theoj.org/reviewers (without tagging them with an @), thanks.
