### 2024-11-21 #7370 @evanhockings
Hi @phibeck! I checked the list, filtering for Julia and quantum computing-adjacent people, and ended up with the following shortlist:

- Hugo Strand
- Peter-Jan Derks
- Utkarsh
- Zhen Huang
- Jinguo Liu
- Aleksandr Berezutskii
