### 2024-12-03 #7370 @phibeck
Hi @evanhockings, thanks for the suggestions. Regarding your question, in principle that's within the spirit of JOSS, i.e. recognizing that software is dynamic and new features will be added after publication. I am not sure we can update the paper after publication, but that's why we only require a summary of the high-level functionality and a statement of need, assuming these will not become outdated. You can always add documentation to describe the new features in the repo directly. What you could do is include a statement in the manuscript or the repo about intended developments, for example. I will ask the editors and get back to you.
