### 2025-01-06 #7370 @phibeck
> Hey @phibeck! Sorry for the late response, I had been on vacation. Let me know if this still requires a review, I could take a look in the coming weeks.

Hi @obliviateandsurrender thanks, that would be great!
