### 2025-01-23 c6fea96
v0.2 update changes
