+++ src/circuits/heavy_hex.jl
@@ -1,0 +1,557 @@
+"""
+    HeavyHexParameters
+
+Parameters for the syndrome extraction circuit of a heavy hex code.
+
+# Fields
+
+  - `params::Dict{Symbol, Any}`: Dictionary of the circuit parameters described below.
+  - `circuit_name::String`: Name of the circuit used for saving data.
+
+# Parameters
+
+  - `vertical_dist::Int`: Vertical (Z) distance of the code.
+  - `horizontal_dist::Int`: Horizontal (X) distance of the code.
+  - `flipped::Bool`: Whether ancilla qubit layout is flipped left to right.
+  - `gate_type::Symbol`: Type of two-qubit gate used in the circuit, which must be `:cx`.
+  - `ancilla_measurement::Bool`: Whether to include mid-circuit ancilla measurements.
+  - `pad_identity::Bool`: Whether to pad layers with single-qubit identity gates.
+  - `layer_time_dict::Dict{Symbol, Float64}`: Dictionary of layer times.
+"""
+struct HeavyHexParameters <: AbstractCircuitParameters
+    params::Dict{Symbol, Any}
+    circuit_name::String
+    # Default constructor
+    function HeavyHexParameters(params::Dict{Symbol, Any}, circuit_name::String)
+        # Check circuit parameters are present
+        @assert haskey(params, :vertical_dist) "The vertical distance is missing."
+        @assert haskey(params, :horizontal_dist) "The horizontal distance is missing."
+        @assert haskey(params, :flipped) "The flipped flag is missing."
+        @assert haskey(params, :gate_type) "The gate type is missing."
+        @assert haskey(params, :ancilla_measurement) "The ancilla measurement flag is missing."
+        @assert haskey(params, :pad_identity) "The pad identity flag is missing."
+        @assert haskey(params, :layer_time_dict) "The layer time dictionary is missing."
+        vertical_dist = params[:vertical_dist]
+        horizontal_dist = params[:horizontal_dist]
+        flipped = params[:flipped]
+        gate_type = params[:gate_type]
+        ancilla_measurement = params[:ancilla_measurement]
+        pad_identity = params[:pad_identity]
+        layer_time_dict = params[:layer_time_dict]
+        # Check some conditions
+        @assert (vertical_dist >= 3 && horizontal_dist >= 3) "Invalid distance $(vertical_dist) x $(horizontal_dist). Must be at least 3 x 3."
+        @assert gate_type == :cx "Invalid gate type $(gate_type). Must be :cx."
+        @assert haskey(layer_time_dict, :single_qubit) "The layer time dictionary must contain the key :single_qubit."
+        @assert haskey(layer_time_dict, :two_qubit) "The layer time dictionary must contain the key :two_qubit."
+        @assert haskey(layer_time_dict, :meas_reset) "The layer time dictionary must contain the key :meas_reset."
+        @assert haskey(layer_time_dict, :mid_reset) "The layer time dictionary must contain the key :mid_reset."
+        @assert layer_time_dict[:single_qubit] > 0.0 "The single-qubit layer time must be positive."
+        @assert layer_time_dict[:two_qubit] > 0.0 "The two-qubit layer time must be positive."
+        @assert layer_time_dict[:meas_reset] > 0.0 "The measurement and reset time must be positive."
+        @assert layer_time_dict[:mid_reset] > 0.0 "The mid-circuit reset layer time must be positive."
+        # Set the circuit name
[…]
