+++ docs/make.jl
@@ -21,14 +21,23 @@ makedocs(;
                 "QuantumACES.jl" => "internal/QuantumACES.md",
                 "tableau.jl" => "internal/tableau.md",
                 "noise.jl" => "internal/noise.md",
+                "noises/depolarising.jl" => "internal/noises/depolarising.md",
+                "noises/lognormal.jl" => "internal/noises/lognormal.md",
                 "circuit.jl" => "internal/circuit.md",
+                "circuits/rotated_planar.jl" => "internal/circuits/rotated_planar.md",
+                "circuits/unrotated_planar.jl" => "internal/circuits/unrotated_planar.md",
+                "circuits/heavy_hex.jl" => "internal/circuits/heavy_hex.md",
+                "stim.jl" => "internal/stim.md",
                 "tuples.jl" => "internal/tuples.md",
                 "design.jl" => "internal/design.md",
+                "rand_design.jl" => "internal/rand_design.md",
                 "merit.jl" => "internal/merit.md",
-                "weights.jl" => "internal/weights.md",
-                "optimise.jl" => "internal/optimise.md",
-                "scaling.jl" => "internal/scaling.md",
+                "optimise_weights.jl" => "internal/optimise_weights.md",
+                "optimise_tuples.jl" => "internal/optimise_tuples.md",
+                "estimate.jl" => "internal/estimate.md",
                 "simulate.jl" => "internal/simulate.md",
+                "device.jl" => "internal/device.md",
+                "scaling.jl" => "internal/scaling.md",
                 "kwargs.jl" => "internal/kwargs.md",
                 "utils.jl" => "internal/utils.md",
                 "io.jl" => "internal/io.md",

[…]
