+++ .JuliaFormatter.toml
@@ -8,3 +8,4 @@ format_docstrings = true
 normalize_line_endings = "unix"
 separate_kwargs_with_semicolon = true
 format_markdown = true
+ignore = "paper/"
