+++ .JuliaFormatter.toml
@@ -7,5 +7,3 @@ whitespace_in_kwargs = true
 format_docstrings = true
 normalize_line_endings = "unix"
 separate_kwargs_with_semicolon = true
-format_markdown = true
-ignore = "paper/"
