+++ docs/src/internal/merit.md
@@ -8,7 +8,7 @@ CollapsedDocStrings = true
 
 ```@autodocs; canonical=false
 Modules = [QuantumACES]
-Pages = ["merit.jl"]
+Pages = ["src/merit.jl"]
 Private = false
 ```
 
@@ -16,6 +16,6 @@ Private = false
 
 ```@autodocs
 Modules = [QuantumACES]
-Pages = ["merit.jl"]
+Pages = ["src/merit.jl"]
 Public = false
 ```
