+++ docs/src/internal/design.md
@@ -8,7 +8,7 @@ CollapsedDocStrings = true
 
 ```@autodocs; canonical=false
 Modules = [QuantumACES]
-Pages = ["design.jl"]
+Pages = ["src/design.jl"]
 Private = false
 ```
 
@@ -16,6 +16,6 @@ Private = false
 
 ```@autodocs
 Modules = [QuantumACES]
-Pages = ["design.jl"]
+Pages = ["src/design.jl"]
 Public = false
 ```
