+++ docs/src/internal/noise.md
@@ -8,7 +8,7 @@ CollapsedDocStrings = true
 
 ```@autodocs; canonical=false
 Modules = [QuantumACES]
-Pages = ["noise.jl"]
+Pages = ["src/noise.jl"]
 Private = false
 ```
 
@@ -16,6 +16,6 @@ Private = false
 
 ```@autodocs
 Modules = [QuantumACES]
-Pages = ["noise.jl"]
+Pages = ["src/noise.jl"]
 Public = false
 ```
