+++ docs/src/internal/QuantumACES.md
@@ -8,7 +8,7 @@ CollapsedDocStrings = true
 
 ```@autodocs; canonical=false
 Modules = [QuantumACES]
-Pages = ["QuantumACES.jl"]
+Pages = ["src/QuantumACES.jl"]
 Private = false
 ```
 
@@ -16,6 +16,6 @@ Private = false
 
 ```@autodocs
 Modules = [QuantumACES]
-Pages = ["QuantumACES.jl"]
+Pages = ["src/QuantumACES.jl"]
 Public = false
 ```
