+++ docs/src/internal/circuit.md
@@ -8,7 +8,7 @@ CollapsedDocStrings = true
 
 ```@autodocs; canonical=false
 Modules = [QuantumACES]
-Pages = ["circuit.jl"]
+Pages = ["src/circuit.jl"]
 Private = false
 ```
 
@@ -16,6 +16,6 @@ Private = false
 
 ```@autodocs
 Modules = [QuantumACES]
-Pages = ["circuit.jl"]
+Pages = ["src/circuit.jl"]
 Public = false
 ```
