+++ docs/src/internal/circuits/unrotated_planar.md
@@ -8,7 +8,7 @@ CollapsedDocStrings = true
 
 ```@autodocs; canonical=false
 Modules = [QuantumACES]
-Pages = ["circuits/unrotated_planar.jl"]
+Pages = ["src/circuits/unrotated_planar.jl"]
 Private = false
 ```
 
@@ -16,6 +16,6 @@ Private = false
 
 ```@autodocs
 Modules = [QuantumACES]
-Pages = ["circuits/unrotated_planar.jl"]
+Pages = ["src/circuits/unrotated_planar.jl"]
 Public = false
 ```
