+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "QuantumACES.jl: Design noise characterisation experiments for quantum computers"
+title: "QuantumACES.jl: design noise characterisation experiments for quantum computers"
 tags:
   - Julia
   - Quantum computing
@@ -13,42 +13,46 @@ affiliations:
     index: 1
   - name: ARC Centre of Excellence for Engineered Quantum Systems
     index: 2
-date: 16 September 2024
+date: 23 January 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-`QuantumACES.jl` is a Julia [@bezanson_julia_2017] package for designing, optimising, and simulating scalable and performant noise characterisation experiments for quantum computers.
-
+QuantumACES.jl is a Julia&nbsp;[@bezanson_julia_2017] package for designing, simulating, and performing scalable Pauli noise characterisation experiments for quantum computers.
 Noise in quantum devices is the key obstacle to large-scale quantum computation.
-Consequently, quantum computers will require fault-tolerant architectures that encode quantum information in quantum error correcting codes [@shor_faulttolerant_1996; @gottesman_introduction_2010].
-Detailed characterisation of the noise in prototype quantum computers can demonstrate that the device is performing correctly, and improve the performance of error correction by identifying which physical configurations of errors are more and less likely [@sundaresan_demonstrating_2023].
+Consequently, quantum computers will require fault-tolerant architectures that replace physical qubits and operations with redundantly-encoded logical equivalents, which entails regularly measuring the parity checks of quantum error correcting codes&nbsp;[@shor_faulttolerant_1996; @gottesman_introduction_2010; @aliferis_introduction_2013].
+Detailed characterisation of the noise in prototype quantum computers can demonstrate correct device performance and identify poorly functioning qubits and operations.
+It can also enable noise-aware decoding of quantum error correcting codes by calibrating the decoder prior on which physical configurations of errors are more and less likely&nbsp;[@tuckett_faulttolerant_2020; @chen_calibrated_2022; @sundaresan_demonstrating_2023; @tiurev_correcting_2023; @higgott_improved_2023].
 
-`QuantumACES.jl` designs experiments with the framework of averaged circuit eigenvalue sampling (ACES) [@flammia_averaged_2022] that characterise Pauli noise in stabiliser circuits, following theory outlined in [@hockings_scalable_2024].
-Stabiliser circuits are a restricted class of quantum circuits that admit efficient classical simulation [@aaronson_improved_2004; @gottesman_stabilizer_1997].
+QuantumACES.jl uses the framework of averaged circuit eigenvalue sampling (ACES)&nbsp;[@flammia_averaged_2022] to design experiments that characterise Pauli noise in stabiliser circuits, following the theory and protocol outlined in&nbsp;[@hockings_scalable_2024].
+Stabiliser circuits are a restricted class of quantum circuits that admit efficient classical simulation&nbsp;[@aaronson_improved_2004; @gottesman_stabilizer_1997], including with Pauli noise.
+Quantum noise is tailored into Pauli noise by techniques such as Pauli frame randomisation&nbsp;[@knill_quantum_2005], randomised compiling&nbsp;[@wallman_noise_2016], or quantum error correction itself&nbsp;[@beale_coherence_2018], and the theory of quantum error correction and fault tolerance generally relies on modelling noise as Pauli noise&nbsp;[@terhal_quantum_2015].
 
-In a typical fault-tolerant quantum computing architecture, syndrome extraction circuits, which are stabiliser circuits that measure the parity checks of quantum error correcting codes, comprise the bulk of the physical qubits and gate operations.
-Moreover, the theory of quantum error correction and fault-tolerance generally relies on modelling noise as Pauli noise [@terhal_quantum_2015].
[…]
