### 2025-01-23 #7370 @evanhockings
Hi @phibeck I've made the changes (including updating the paper) and released the changes as v0.2.0, so everything is ready for review!
