### 2025-01-23 #7707 @evanhockings
Wow, how can I get this reference check tool? Would be useful for other papers! I'll update the bibliography with non-arXiv citations for those two papers.
