### 2025-01-27 a8717a2
Updated padding handling
