+++ paper/paper.bib
@@ -339,7 +339,7 @@
 @inproceedings{shor_faulttolerant_1996,
   title = {Fault-Tolerant Quantum Computation},
   booktitle = {Proceedings of the 37th Annual Symposium on {{Foundations}} of {{Computer Science}}},
-  author = {Shor, P.W.},
+  author = {Shor, Peter W.},
   year = {1996},
   month = oct,
   eprint = {quant-ph/9605011},
