+++ paper/paper.bib
@@ -22,7 +22,7 @@
   year = {2013},
   pages = {126--160},
   publisher = {Cambridge University Press},
-  address = {Cambridge},
+  address = {Cambridge, United Kingdom},
   doi = {10.1017/CBO9781139034807.007},
   url = {https://www.cambridge.org/core/books/quantum-error-correction/introduction-to-quantum-fault-tolerance/21FAD6E04EFAA68FEF5C217BC361EDD8},
   isbn = {978-0-521-89787-7}
@@ -193,7 +193,7 @@
   eprint = {0904.2557},
   pages = {13--58},
   publisher = {American Mathematical Society},
-  address = {USA},
+  address = {Washington, DC, USA},
   doi = {10.1090/psapm/068},
   url = {http://arxiv.org/abs/0904.2557},
   archiveprefix = {arXiv}
@@ -338,12 +338,14 @@
 
 @inproceedings{shor_faulttolerant_1996,
   title = {Fault-Tolerant Quantum Computation},
-  booktitle = {Proceedings of the 37th Annual Symposium on {{Foundations}} of {{Computer Science}}},
+  booktitle = {Proceedings of the 37th {{Annual Symposium}} on {{Foundations}} of {{Computer Science}}},
   author = {Shor, Peter W.},
   year = {1996},
   month = oct,
   eprint = {quant-ph/9605011},
   pages = {56--65},
+  publisher = {IEEE},
+  address = {Burlington, VT, USA},
   doi = {10.1109/SFCS.1996.548464},
   url = {http://arxiv.org/abs/quant-ph/9605011},
   archiveprefix = {arXiv}
