+++ paper/paper.bib
@@ -193,7 +193,7 @@
   eprint = {0904.2557},
   pages = {13--58},
   publisher = {American Mathematical Society},
-  address = {Washington, DC, USA},
+  address = {Providence, RI, USA},
   doi = {10.1090/psapm/068},
   url = {http://arxiv.org/abs/0904.2557},
   archiveprefix = {arXiv}
