### 2025-02-13 #7707 @phibeck
Hi @obliviateandsurrender and @pkairys, just checking in whether you have any questions regarding the review process or if there's anything else that's holding you up, thanks.

@meandmytram I see you've completed your checklist, thank you! 🎉 Could you please let us know whether you have any comments for the authors (I don't see any issues on the repo)? If not, could you please confirm that you completed your checklist and agree to the publication of this software in its current form?
