+++ paper/paper.bib
@@ -142,7 +142,8 @@
   volume = {232},
   eprint = {2108.05803},
   pages = {4:1-4:10},
-  publisher = {Schloss Dagstuhl -- Leibniz-Zentrum f{\"u}r Informatik},
+  publisher = {Schloss Dagstuhl},
+  address = {Urbana-Champaign, IL, USA},
   doi = {10.4230/LIPIcs.TQC.2022.4},
   url = {http://arxiv.org/abs/2108.05803},
   archiveprefix = {arXiv},
