### 2025-02-17 a6dcbd6
Minor paper updates
