+++ paper/paper.bib
@@ -28,6 +28,23 @@
   isbn = {978-0-521-89787-7}
 }
 
+@article{bausch_learning_2024,
+  title = {Learning High-Accuracy Error Decoding for Quantum Processors},
+  author = {Bausch, Johannes and Senior, Andrew W. and Heras, Francisco J. H. and Edlich, Thomas and Davies, Alex and Newman, Michael and Jones, Cody and Satzinger, Kevin and Niu, Murphy Yuezhen and Blackwell, Sam and Holland, George and Kafri, Dvir and Atalaya, Juan and Gidney, Craig and Hassabis, Demis and Boixo, Sergio and Neven, Hartmut and Kohli, Pushmeet},
+  year = {2024},
+  month = nov,
+  journal = {Nature},
+  volume = {635},
+  number = {8040},
+  eprint = {2310.05900},
+  pages = {834--840},
+  issn = {1476-4687},
+  doi = {10.1038/s41586-024-08148-8},
+  url = {https://arxiv.org/abs/2310.05900},
+  archiveprefix = {arXiv},
+  langid = {english}
+}
+
 @article{beale_coherence_2018,
   title = {Coherence in Quantum Error-Correcting Codes},
   author = {Beale, Stefanie J. and Wallman, Joel J. and Guti{\'e}rrez, Mauricio and Brown, Kenneth R. and Laflamme, Raymond},
