### 2025-02-18 #7707 @pkairys
@evanhockings The only thing I don't see is "Community Guidelines" in the sense of the checklist above. I think it's obvious that you want people to contribute via github but it might be useful to add 2-3 sentences along contributing guidelines and points of contact for questions. I'd recommend adding this towards the bottom of the readme or even as a separate tab in the documentation.
