### 2025-02-19 #7707 @evanhockings
@pkairys Cheers, I made a CONTRIBUTING.md file in root directory! It's in the joss-updates branch which also has some minor edits to tighten up the wording of the paper—I'll merge it once the review is done.
