+++ paper/paper.bib
@@ -276,13 +276,19 @@
   langid = {british}
 }
 
-@article{hockings_scalable_2024,
+@article{hockings_scalable_2025,
   title = {Scalable Noise Characterisation of Syndrome Extraction Circuits with Averaged Circuit Eigenvalue Sampling},
   author = {Hockings, Evan T. and Doherty, Andrew C. and Harper, Robin},
-  year = {2024},
-  month = apr,
-  journal = {arXiv preprint},
+  year = {2025},
+  month = feb,
+  journal = {PRX Quantum},
+  volume = {6},
+  number = {1},
   eprint = {2404.06545},
+  pages = {010334},
+  issn = {2691-3399},
+  doi = {10.1103/PRXQuantum.6.010334},
+  url = {https://arxiv.org/abs/2404.06545},
   archiveprefix = {arXiv}
 }
 
