+++ paper/paper.md
@@ -13,44 +13,45 @@ affiliations:
     index: 1
   - name: ARC Centre of Excellence for Engineered Quantum Systems
     index: 2
-date: 23 February 2025
+date: 28 February 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-QuantumACES is a Julia&nbsp;[@bezanson_julia_2017] package for designing, simulating, and performing scalable Pauli noise characterisation experiments for quantum computers.
+QuantumACES is a Julia [@bezanson_julia_2017] package for designing, simulating, and performing scalable Pauli noise characterisation experiments for quantum computers.
 Noise in quantum devices is the key obstacle to large-scale quantum computation.
-Consequently, quantum computers will require fault-tolerant architectures that replace physical qubits and operations with redundantly-encoded logical equivalents, which entails regularly measuring the parity checks of quantum error correcting codes&nbsp;[@shor_faulttolerant_1996; @gottesman_introduction_2010; @aliferis_introduction_2013].
-Decoders attempt to underlying physical errors in the quantum device from the resulting error syndrome data to determine a correction operation that preserves the logical information.
-The noise estimates produced by QuantumACES can calibrate decoders to the likelihood of physical error configurations in a quantum device&nbsp;[@tuckett_faulttolerant_2020; @chen_calibrated_2022; @sundaresan_demonstrating_2023; @tiurev_correcting_2023; @higgott_improved_2023], enabling noise-aware decoding.
-They can also be used to generate simulated data for training more accurate decoders, such as&nbsp;[@bausch_learning_2024], verify appropriate device calibration, and inform the co-design of quantum error correcting codes, decoders, fault-tolerant circuits, and quantum devices.
+Consequently, quantum computers will require fault-tolerant architectures that replace physical qubits and operations with redundantly-encoded logical equivalents, which entails regularly measuring the parity checks of quantum error correcting codes [@shor_faulttolerant_1996; @gottesman_introduction_2010; @aliferis_introduction_2013].
+Decoders process the resulting error syndrome data and attempt to infer the most likely underlying physical errors in the quantum device.
+Then, they determine correction operations that attempt to preserve logical information.
+The noise estimates produced by QuantumACES can inform decoders of the likelihood of physical error configurations in a quantum device [@tuckett_faulttolerant_2020; @chen_calibrated_2022; @sundaresan_demonstrating_2023; @tiurev_correcting_2023; @higgott_improved_2023], enabling noise-aware decoding.
+They can also be used to generate simulated data for training more accurate decoders, such as [@bausch_learning_2024], verify appropriate device calibration, and inform the co-design of quantum error correcting codes, decoders, fault-tolerant circuits, and quantum devices.
 
-QuantumACES designs experiments to characterise Pauli noise in stabiliser circuits within the framework of averaged circuit eigenvalue sampling (ACES)&nbsp;[@flammia_averaged_2022], following the theory and protocol outlined in&nbsp;[@hockings_scalable_2025].
-Stabiliser circuits are a restricted class of quantum circuits that admit efficient classical simulation&nbsp;[@aaronson_improved_2004; @gottesman_stabilizer_1997], including with Pauli noise.
-Quantum noise is tailored into Pauli noise by techniques such as Pauli frame randomisation&nbsp;[@knill_quantum_2005], randomised compiling&nbsp;[@wallman_noise_2016], or quantum error correction itself&nbsp;[@beale_coherence_2018], and the theory of quantum error correction and fault tolerance generally relies on modelling noise as Pauli noise&nbsp;[@terhal_quantum_2015].
+QuantumACES designs experiments to characterise Pauli noise in stabiliser circuits within the framework of averaged circuit eigenvalue sampling (ACES) [@flammia_averaged_2022], following the theory and protocol outlined in [@hockings_scalable_2025].
[…]
