+++ paper/paper.bib
@@ -276,6 +276,15 @@
   langid = {british}
 }
 
+@article{hockings_improving_2025,
+  title = {Improving Error Suppression with Noise-Aware Decoding},
+  author = {Hockings, Evan T. and Doherty, Andrew C. and Harper, Robin},
+  year = {2025},
+  month = feb,
+  eprint = {2502.21044},
+  archiveprefix = {arXiv}
+}
+
 @article{hockings_scalable_2025,
   title = {Scalable Noise Characterisation of Syndrome Extraction Circuits with Averaged Circuit Eigenvalue Sampling},
   author = {Hockings, Evan T. and Doherty, Andrew C. and Harper, Robin},
