+++ paper/paper.md
@@ -13,7 +13,7 @@ affiliations:
     index: 1
   - name: ARC Centre of Excellence for Engineered Quantum Systems
     index: 2
-date: 28 February 2025
+date: 3 March 2025
 bibliography: paper.bib
 ---
 
@@ -23,13 +23,14 @@ QuantumACES is a Julia [@bezanson_julia_2017] package for designing, simulating,
 Noise in quantum devices is the key obstacle to large-scale quantum computation.
 Consequently, quantum computers will require fault-tolerant architectures that replace physical qubits and operations with redundantly-encoded logical equivalents, which entails regularly measuring the parity checks of quantum error correcting codes [@shor_faulttolerant_1996; @gottesman_introduction_2010; @aliferis_introduction_2013].
 Decoders process the resulting error syndrome data and attempt to infer the most likely underlying physical errors in the quantum device.
-Then, they determine correction operations that attempt to preserve logical information.
+Subsequently, they determine correction operations that attempt to preserve logical information.
 The noise estimates produced by QuantumACES can inform decoders of the likelihood of physical error configurations in a quantum device [@tuckett_faulttolerant_2020; @chen_calibrated_2022; @sundaresan_demonstrating_2023; @tiurev_correcting_2023; @higgott_improved_2023], enabling noise-aware decoding.
 They can also be used to generate simulated data for training more accurate decoders, such as [@bausch_learning_2024], verify appropriate device calibration, and inform the co-design of quantum error correcting codes, decoders, fault-tolerant circuits, and quantum devices.
 
 QuantumACES designs experiments to characterise Pauli noise in stabiliser circuits within the framework of averaged circuit eigenvalue sampling (ACES) [@flammia_averaged_2022], following the theory and protocol outlined in [@hockings_scalable_2025].
 Stabiliser circuits are a restricted class of quantum circuits that admit efficient classical simulation [@aaronson_improved_2004; @gottesman_stabilizer_1997], including with Pauli noise.
-Quantum noise is tailored into Pauli noise by techniques such as Pauli frame randomisation [@knill_quantum_2005], randomised compiling [@wallman_noise_2016], or quantum error correction itself [@beale_coherence_2018], and the theory of quantum error correction and fault tolerance generally relies on modelling noise as Pauli noise [@terhal_quantum_2015].
+Quantum noise is tailored into Pauli noise by techniques such as Pauli frame randomisation [@knill_quantum_2005], randomised compiling [@wallman_noise_2016], or quantum error correction itself [@beale_coherence_2018].
+Additionally, the theory of quantum error correction and fault tolerance generally relies on modelling noise as Pauli noise [@terhal_quantum_2015].
 
 QuantumACES contains routines for optimising designs for noise characterisation experiments, given an arbitrary stabiliser circuit and Pauli noise model, using functions that precisely predict the performance of these experimental designs.
 It has built-in circuits and noise models and also allows users to define their own.
@@ -38,16 +39,17 @@ These noise characterisation experiments are simulated with the open-source Pyth
 In a typical fault-tolerant quantum computing architecture, the bulk of the physical qubits and gate operations are dedicated to performing the syndrome extraction circuits that measure the parity checks of quantum error correcting codes.
 These syndrome extraction circuits, which are stabiliser circuits, are therefore the key target for noise characterisation experiments.
 QuantumACES is tailored to characterising Pauli noise in syndrome extraction circuits, particularly for topological quantum error correcting codes such as the surface code [@bravyi_quantum_1998; @dennis_topological_2002; @kitaev_faulttolerant_2003; @fowler_surface_2012].
[…]
