+++ docs/example.jl
@@ -19,56 +19,71 @@ display(d)
 d = optimise_design(circuit_dep; options = OptimOptions(; seed = seed))
 display(d)
 # Simulate repeated noise characterisation experiments
+d_log = generate_design(circuit_log, d)
 d_log = update_noise(d, log_param)
 budget_set = [10^6; 10^7; 10^8]
 repetitions = 20
 aces_data = simulate_aces(d_log, budget_set; repetitions = repetitions, seed = seed)
 # Compare the performance to predictions
 merit_log = calc_merit(d_log)
-display(merit_log)
 pretty_print(aces_data, merit_log)
-noise_score_coll = get_noise_score(aces_data, merit_log)
-gls_z_scores = [noise_score.gls_z_score for noise_score in noise_score_coll]
-display(gls_z_scores)
-# Examine the performance scaling
-dist_max = 7
-merit_scaling = calc_merit_scaling(d, dist_max)
-ensemble_scaling = calc_ensemble_scaling(d_log, dist_max; seed = seed)
 # Generate a design for a large circuit
 dist_big = 13
 rotated_param_big = get_rotated_param(dist_big)
-circuit_big = get_circuit(rotated_param_big, log_param)
-d_big = generate_design(
-    circuit_big,
-    d.tuple_set_data;
-    shot_weights = d.shot_weights,
-    full_covariance = false,
-    diagnostics = true,
-)
+circuit_big = get_circuit(rotated_param_big, dep_param)
+circuit_big_log = get_circuit(rotated_param_big, log_param)
+d_big = generate_design(circuit_big_log, d; full_covariance = false, diagnostics = true)
 # Simulate a large noise characterisation experiment
 aces_data_big = simulate_aces(d_big, budget_set; seed = seed, split = true)
-# Compare the performance to predictions
+# Examine the performance scaling
+dist_max = 7
+merit_scaling = calc_merit_scaling(d, dist_max)
 scaling_fit = get_scaling_fit(merit_scaling)
+ensemble_scaling = calc_ensemble_scaling(d_log, dist_max; seed = seed)
 ensemble_fit = get_ensemble_fit(ensemble_scaling)
+# Compare the performance to predictions
 wls_pred_expectation = ensemble_fit.wls_expectation_model(dist_big)
 wls_pred_variance = ensemble_fit.wls_variance_model(dist_big)
-wls_z_scores_big = [
+wls_scores_big = [
     (noise_error.wls_nrmse .- wls_pred_expectation) / sqrt(wls_pred_variance) for
     noise_error in aces_data_big.noise_error_coll[1, :]
 ]
-display(wls_z_scores_big)
+display(wls_scores_big)
 wls_pred_relative_expectation = ensemble_fit.wls_relative_expectation_model(dist_big)
 wls_pred_relative_variance = ensemble_fit.wls_relative_variance_model(dist_big)
-wls_relative_z_scores_big = [
+wls_relative_scores_big = [
     (noise_error.wls_relative_nrmse .- wls_pred_relative_expectation) /
     sqrt(wls_pred_relative_variance) for
     noise_error in aces_data_big.noise_error_coll[1, :]
 ]
-display(wls_relative_z_scores_big)
+display(wls_relative_scores_big)
+# Simulate a memory experiment
+big_rounds = dist_big
+big_shots = 5 * 10^6
+decoder_gate_probabilities = [
+    circuit_big_log.gate_probabilities
+    circuit_big.gate_probabilities
+    [noise_est.wls_gate_probabilities for noise_est in aces_data_big.noise_est_coll[1, :]]
+]
+decoder_labels = [
+    "True"
+    "Depolarising"
[…]
