### 2025-03-03 #7707 @evanhockings
Okay, I made a few updates to the paper and have now merged the branch into main as part of a new release. Everything's good to go on my end.
