### 2025-03-10 #7707 @obliviateandsurrender
Sorry for the delay in reviewing the work! Finally got a weekend to take a look. Overall, really great work! Everything looks publishable at its current stage, just putting down some suggestion after which I can give my approval:

Suggestions for the library
1. The link to [OptimOptions](https://github.com/evanhockings/QuantumACES.jl/blob/main/@ref) in the README seems to be broken.
2. Could we mention a couple of keywords for reducing the runtime of the `optimise_design` in the README? In fact, it might be nice to have at least one example workflow which is quicker to run in addition to the current one.
3. In the docs, under the Guide, `"Package Usage"` can be changed to something more intuitive. In fact, repackaging the current examples as a separate section should help a lot with the usability of the current docs.

Suggestions for the article
1. redundantly-encoded --> redundantly encoded
2. they determine correction operations --> they help determine correction operations that attempt...
3. such as (Bausch et al. 2024) --> such as one in (Bausch et al. 2024)
4. "verify appropriate device calibrations"..., --> "to verify appropriate device calibration, and to inform the co-design..."
5. To make the statement of need more readable, we can skip defining the acronyms that are not used.
