+++ docs/src/creation.md
@@ -1,0 +1,353 @@
+# Creating Circuits and Noise Models
+
+`QuantumACES` makes it easy to create new circuits and noise models.
+At a high level, we create new parameter types for the circuit or noise model and then create new methods for the functions [`get_circuit`](@ref) and [`init_gate_probabilities`](@ref), respectively, that take these parameter types as arguments.
+Then [`get_circuit`](@ref) uses the circuit and noise model parameters to create a [`Circuit`](@ref) object.
+
+The [`Circuit`](@ref) object enables the functionality we saw in [Package usage](@ref), with two exceptions.
+First, calculating performance predictions requires the circuit to be parameterised by a `dist` parameter, typically the distance of the code underlying the syndrome extraction circuit.
+Second, simulating memory experiments in Stim requires the circuit to be a syndrome extraction annotated with the appropriate information.
+The [`Circuit`](@ref) object contains an `extra_fields` field which is a dictionary that can store additional parameters to enable functionality such as this.
+In particular, syndrome extraction circuits must store a [`CodeParameters`](@ref)object in this dictionary, which is then used to generate the detectors in Stim for memory experiment circuits, enabling decoding.
+
+The example code shown below can be found [here](../example/creation.jl).
+
+Let us begin by creating a new circuit, following the example circuit shown in Figure 2 of [arXiv:2404.06545](https://arxiv.org/abs/2404.06545).
+The first step is to create a parameter object for the circuit, which must be a subtype of [`AbstractCircuitParameters`](@ref) and contain the necessary fields `params` and `circuit_name`, using the `StructEquality` package to automatically generate hash and equality relationships for the object to enable comparisons for the resulting [`Circuit`](@ref) objects.
+
+```julia
+using QuantumACES, StructEquality
+struct ExampleParameters <: AbstractCircuitParameters
+    params::Dict{Symbol, Any}
+    circuit_name::String
+    # Default constructor
+    function ExampleParameters(params::Dict{Symbol, Any}, circuit_name::String)
+        # Check circuit parameters are present
+        @assert haskey(params, :pad_identity) "The pad identity flag is missing."
+        @assert haskey(params, :layer_time_dict) "The layer time dictionary is missing."
+        pad_identity = params[:pad_identity]
+        layer_time_dict = params[:layer_time_dict]
+        # Check some conditions
+        @assert haskey(layer_time_dict, :single_qubit) "The layer time dictionary must contain the key :single_qubit."
+        @assert haskey(layer_time_dict, :two_qubit) "The layer time dictionary must contain the key :two_qubit."
+        @assert haskey(layer_time_dict, :meas_reset) "The layer time dictionary must contain the key :meas_reset."
+        @assert layer_time_dict[:single_qubit] > 0.0 "The single-qubit layer time must be positive."
+        @assert layer_time_dict[:two_qubit] > 0.0 "The two-qubit layer time must be positive."
+        @assert layer_time_dict[:meas_reset] > 0.0 "The measurement and reset time must be positive."
+        # Set the circuit name
+        new_circuit_name = "example_circuit"
+        if pad_identity != true
+            circuit_name *= "_no_pad_identity"
+        end
+        # Return parameters
+        return new(params, new_circuit_name)::ExampleParameters
+    end
+end
+
+@struct_hash_equal_isequal ExampleParameters
+```
+
+We need a function to construct the parameter object.
+
+```julia
+function get_example_param(;
+    pad_identity = true,
[…]
