+++ docs/src/performance.md
@@ -39,4 +39,4 @@ Once we have the gate log-eigenvalues, we can straightforwardly estimate the Pau
 However, these estimates are not guaranteed to be valid probability distributions, so we must project the estimates into the probability simplex.
 As discussed in [arXiv:2502.21044](https://arxiv.org/abs/2502.21044), we perform this projection in the Mahalanobis distance with the fast convex solver [SCS.jl](https://github.com/jump-dev/SCS.jl).
 We can optimise calculation of the precision matrix such that the only matrix inversion required is the highly-optimised inversion of the circuit log-eigenvalue estimator covariance matrix ``\Omega^\prime``.
-This projection performs best, at least at small scales, when single-threaded by setting `ENV["OMP_NUM_THREADS"]="1"` in `~/.julia/config/startup.jl`, but this causes large performance regressions in other linear algebra routines in the package, so the setting is only advised when benchmarking the time taken to perform ACES noise estimation.
+This projection performs best, at least at small scales, when single-threaded by setting `ENV["OMP_NUM_THREADS"] = "1"` in `~/.julia/config/startup.jl`, but this causes large performance regressions in other linear algebra routines in the package, so the setting is only advised when benchmarking the time taken to perform ACES noise estimation.

[…]
