### 2025-03-11 #7707 @evanhockings
Hi @obliviateandsurrender, thanks for the suggestions! I've implemented them as follows:

Library:

1. Ah, link removed! I copied that from the package guide, and it doesn't work in the README as it's not in the docs.
2. I added the main options you'd want to toggle.
3. I split up the guide between the main examples (which are now under 'Example usage') and the circuit and noise model creation examples. I also point to the circuit and noise model creation examples as being faster to run, in response to 2.

Article:

1, 3, 4, 5. Changed.

2. Adding 'help' feels like it changes the meaning slightly.

I'll create a new release for the package once this paper is published!
