+++ docs/src/public.md
@@ -1,197 +1,0 @@
-```@meta
-CollapsedDocStrings = true
-```
-
-# Public API
-
-Documentation for the public API of `QuantumACES`.
-
-To use the internal API, which is documented in the internal API section, it may be convenient to import the module by running the following.
-
-```julia
-import QuantumACES as ACES
-```
-
-## QuantumACES.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/QuantumACES.jl"]
-Private = false
-```
-
-## tableau.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/tableau.jl"]
-Private = false
-```
-
-## noise.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/noise.jl"]
-Private = false
-```
-
-## noises/depolarising.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/noises/depolarising.jl"]
-Private = false
-```
-
-## noises/lognormal.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/noises/lognormal.jl"]
-Private = false
-```
-
-## circuit.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/circuit.jl"]
-Private = false
-```
-
-## circuits/rotated_planar.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/circuits/rotated_planar.jl"]
-Private = false
-```
-
-## circuits/unrotated_planar.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/circuits/unrotated_planar.jl"]
-Private = false
-```
-
-## circuits/heavy_hex.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/circuits/heavy_hex.jl"]
-Private = false
-```
-
-## stim.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/stim.jl"]
-Private = false
-```
-
-## tuples.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/tuples.jl"]
-Private = false
-```
-
-## design.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/design.jl"]
-Private = false
-```
-
-## rand_design.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/rand_design.jl"]
-Private = false
-```
-
-## merit.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/merit.jl"]
-Private = false
-```
-
-## optimise_weights.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/optimise_weights.jl"]
-Private = false
-```
-
-## optimise_tuples.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/optimise_tuples.jl"]
-Private = false
-```
-
-## estimate.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/estimate.jl"]
-Private = false
-```
-
-## simulate.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/simulate.jl"]
-Private = false
-```
-
-## device.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/device.jl"]
-Private = false
-```
-
-## scaling.jl
-
-```@autodocs
-Modules = [QuantumACES]
-Pages = ["src/scaling.jl"]
-Private = false
-```
-
[…]
