+++ docs/src/internal/circuits/rotated_planar.md
@@ -6,7 +6,7 @@ CollapsedDocStrings = true
 
 ## Public functions
 
-```@autodocs; canonical=false
+```@autodocs
 Modules = [QuantumACES]
 Pages = ["src/circuits/rotated_planar.jl"]
 Private = false
