### 2025-03-18 #7707 @evanhockings
Okay @phibeck I've released v0.2.2 and archived it on Zenodo. Here's the badge:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15043208.svg)](https://doi.org/10.5281/zenodo.15043208)

And this is the URL:

https://doi.org/10.5281/zenodo.15043208
