### 2025-03-20 #7707 @phibeck
Hi @evanhockings, could you please adjust the archive title to match that of the JOSS paper? Thanks
