### 2025-03-21 #7707 @phibeck
@evanhockings thanks! I have only two comments regarding the manuscript text:
- lines 4f: in your affiliation, please spell out NSW
- lines 4f: are these two separate affiliations or is it one? I saw in your other papers for example this was declared as a single reference: ARC Centre of Excellence for Engineered Quantum Systems, School of Physics, The University of Sydney, Sydney, New South Wale 2006, Australia. Which one is correct?
