+++ paper/paper.md
@@ -7,10 +7,7 @@ tags:
 authors:
   - name: Evan T. Hockings
     orcid: 0000-0002-5890-5705
-    affiliation: 1
-affiliations:
-  - name: Australian Research Council (ARC) Centre of Excellence for Engineered Quantum Systems, School of Physics, The University of Sydney, Sydney, New South Wales 2006, Australia
-    index: 1
+    affiliation: Australian Research Council (ARC) Centre of Excellence for Engineered Quantum Systems, School of Physics, The University of Sydney, Sydney, New South Wales 2006, Australia
 date: 22 March 2025
 bibliography: paper.bib
 ---
