### 2025-03-26 a2d10be
Figure style updates
