+++ scalable_aces/figures/expectation_heatmap_unrotated_planar_3_3_lognormal.tex
@@ -1,0 +1,901 @@
+\RequirePackage{luatex85}
+\documentclass[tikz]{standalone}
+% Default preamble
+\usepackage{pgfplots}
+\pgfplotsset{compat=newest}
+\usepgfplotslibrary{groupplots}
+\usepgfplotslibrary{polar}
+\usepgfplotslibrary{smithchart}
+\usepgfplotslibrary{statistics}
+\usepgfplotslibrary{dateplot}
+\usepgfplotslibrary{ternary}
+% Custom preamble from global variable:
+\usepackage{amsmath}
+\usepackage{amssymb}
+\usetikzlibrary{arrows.meta}
+\usetikzlibrary{backgrounds}
+\usepgfplotslibrary{patchplots}
+\usepgfplotslibrary{fillbetween}
+\pgfplotsset{%
+    layers/standard/.define layer set={%
+        background,axis background,axis grid,axis ticks,axis lines,axis tick labels,pre main,main,axis descriptions,axis foreground%
+    }{
+        grid style={/pgfplots/on layer=axis grid},%
+        tick style={/pgfplots/on layer=axis ticks},%
+        axis line style={/pgfplots/on layer=axis lines},%
+        label style={/pgfplots/on layer=axis descriptions},%
+        legend style={/pgfplots/on layer=axis descriptions},%
+        title style={/pgfplots/on layer=axis descriptions},%
+        colorbar style={/pgfplots/on layer=axis descriptions},%
+        ticklabel style={/pgfplots/on layer=axis tick labels},%
+        axis background@ style={/pgfplots/on layer=axis background},%
+        3d box foreground style={/pgfplots/on layer=axis foreground},%
+    },
+}
+
+\pgfplotsset{
+colormap={plots1}{rgb(0.00000000)=(0.94634709,0.92901013,0.92575324)
+rgb(0.00392157)=(0.94371151,0.92446250,0.92067015)
+rgb(0.00784314)=(0.94132632,0.91984736,0.91542410)
+rgb(0.01176471)=(0.93905471,0.91520940,0.91010904)
+rgb(0.01568627)=(0.93686346,0.91055922,0.90474739)
+rgb(0.01960784)=(0.93473589,0.90590202,0.89935023)
+rgb(0.02352941)=(0.93266145,0.90124101,0.89392449)
+rgb(0.02745098)=(0.93063333,0.89657817,0.88847456)
+rgb(0.03137255)=(0.92864616,0.89191502,0.88300387)
+rgb(0.03529412)=(0.92669605,0.88725260,0.87751487)
+rgb(0.03921569)=(0.92477958,0.88259180,0.87200971)
+rgb(0.04313725)=(0.92289448,0.87793315,0.86648973)
+rgb(0.04705882)=(0.92103803,0.87327730,0.86095667)
+rgb(0.05098039)=(0.91920879,0.86862452,0.85541133)
+rgb(0.05490196)=(0.91740482,0.86397523,0.84985489)
+rgb(0.05882353)=(0.91562445,0.85932976,0.84428840)
+rgb(0.06274510)=(0.91386714,0.85468808,0.83871202)
+rgb(0.06666667)=(0.91213057,0.85005074,0.83312732)
[…]
