### 2025-03-26 1240d6d
Figure style updates
